CacheMgt refactor
https://github.com/metasfresh/metasfresh/issues/4177

+++ de.metas.inbound.mail/src/main/java/de/metas/inbound/mail/config/InboundEMailConfigChangesDispatcher.java
@@ -56,7 +56,7 @@ public class InboundEMailConfigChangesDispatcher implements ICacheResetListener
 	}
 
 	@Override
-	public int reset(final CacheInvalidateMultiRequest multiRequest)
+	public long reset(final CacheInvalidateMultiRequest multiRequest)
 	{
 		if (listeners.isEmpty())
 		{
